pyproject.toml
[project]
name = "llm-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.4",
    "pypdf2>=3.0.1",
]
